+++ public/admin/_app/immutable/nodes/24.DX8yEQ6u.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/HY4fHbXn.js";export{e as component};